refactor: 리스트 컴포넌트 내에서 쿼리사용
