+++ src/app/real-estate/[id]/page.tsx
@@ -5,8 +5,7 @@ import { Header } from "@/layout/Header";
 import { useRouter } from "next/navigation";
 import RealEstateCallButton from "@/components/common/RealEstateCallButton";
 import { useRealEstateInfoQuery } from "@/queries/real-estate/useRealEstateInfoQuery";
-import { useRealEstatePropertiesQuery } from "@/queries/real-estate/useRealEstatePropertiesQuery";
-import { use, useState, useRef } from "react";
+import { use, useState } from "react";
 
 const statsItems = [
   { label: "월세", value: "rent" },
@@ -19,19 +18,11 @@ const phoneNumbers = [
   { label: "cellPhoneNo", value: "010-1234-5678" },
 ];
 
-// 탭 값과 tradeType 매핑
-const tabToTradeType = {
-  rent: "월세",
-  lease: "전세",
-  deal: "매매",
-};
-
 const RealEstatePage = ({ params }: { params: Promise<{ id: string }> }) => {
   const router = useRouter();
   const { id } = use(params);
   const realtyId = parseInt(id);
   const [selectedTab, setSelectedTab] = useState("rent"); // 기본값은 월세
-  const emptyRef = useRef<HTMLDivElement>(null);
 
   // 부동산 정보
   const {
@@ -39,18 +30,6 @@ const RealEstatePage = ({ params }: { params: Promise<{ id: string }> }) => {
     isLoading: isInfoLoading,
     error: infoError,
   } = useRealEstateInfoQuery(realtyId, { realtyId }, !!realtyId);
-
-  // 선택된 탭에 따른 매물 조회
-  const {
-    items: propertiesData,
-    loader: propertiesLoader,
-    hasMore: propertiesHasMore,
-    loading: isPropertiesLoading,
-  } = useRealEstatePropertiesQuery(
-    realtyId,
-    tabToTradeType[selectedTab as keyof typeof tabToTradeType],
-    20,
-  );
 
   if (isInfoLoading) {
     return (
@@ -85,12 +64,6 @@ const RealEstatePage = ({ params }: { params: Promise<{ id: string }> }) => {
     statsItems,
   };
 
-  // 매물 데이터 추출
-  const properties =
-    propertiesData?.map((property) => ({
-      ...property,
-    })) || [];
-
   return (
     <>
       <Header>
@@ -109,26 +82,8 @@ const RealEstatePage = ({ params }: { params: Promise<{ id: string }> }) => {
             lease: realEstateData.leaseCount,
             deal: realEstateData.dealCount,
           }}
-          propertyMap={{
-            rent: selectedTab === "rent" ? properties : [],
-            lease: selectedTab === "lease" ? properties : [],
-            deal: selectedTab === "deal" ? properties : [],
-          }}
-          loaders={{
-            rent: selectedTab === "rent" ? propertiesLoader : emptyRef,
-            lease: selectedTab === "lease" ? propertiesLoader : emptyRef,
-            deal: selectedTab === "deal" ? propertiesLoader : emptyRef,
-          }}
-          hasMore={{
-            rent: selectedTab === "rent" ? propertiesHasMore : false,
-            lease: selectedTab === "lease" ? propertiesHasMore : false,
-            deal: selectedTab === "deal" ? propertiesHasMore : false,
-          }}
-          loading={{
-            rent: selectedTab === "rent" ? isPropertiesLoading : false,
-            lease: selectedTab === "lease" ? isPropertiesLoading : false,
-            deal: selectedTab === "deal" ? isPropertiesLoading : false,
-          }}
+          realtyId={realtyId}
+          isRealEstatePage={true}
         />
       </div>
       <RealEstateCallButton phoneNumber={phoneNumbers} />
